+++ noodles-bam/src/reader/record.rs
@@ -11,7 +11,7 @@ use std::{
 use byteorder::{LittleEndian, ReadBytesExt};
 use bytes::{Buf, BytesMut};
 use noodles_core::Position;
-use noodles_sam as sam;
+use noodles_sam::{self as sam, reader::record::Fields};
 
 use crate::Record;
 
@@ -37,40 +37,124 @@ where
     Ok(block_size)
 }
 
+pub(super) fn read_record_with_fields<R>(
+    reader: &mut R,
+    buf: &mut BytesMut,
+    record: &mut Record,
+    fields: Fields,
+) -> io::Result<usize>
+where
+    R: Read,
+{
+    let block_size = match reader.read_u32::<LittleEndian>() {
+        Ok(bs) => usize::try_from(bs).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?,
+        Err(ref e) if e.kind() == io::ErrorKind::UnexpectedEof => return Ok(0),
+        Err(e) => return Err(e),
+    };
+
+    buf.resize(block_size, Default::default());
+    reader.read_exact(buf)?;
+
+    decode_record_with_fields(buf, record, fields)?;
+
+    Ok(block_size)
+}
+
 pub(crate) fn decode_record(src: &mut BytesMut, record: &mut Record) -> io::Result<()> {
+    decode_record_with_fields(src, record, Fields::all())
+}
+
+fn decode_record_with_fields(
+    src: &mut BytesMut,
+    record: &mut Record,
+    fields: Fields,
+) -> io::Result<()> {
     use super::alignment_record::{
         get_cigar, get_data, get_mapping_quality, get_quality_scores, get_read_name, get_sequence,
     };
 
-    *record.reference_sequence_id_mut() = get_reference_sequence_id(src)?;
-    *record.position_mut() = get_position(src)?;
+    if fields.contains(Fields::REFERENCE_SEQUENCE_ID) {
+        *record.reference_sequence_id_mut() = get_reference_sequence_id(src)?;
+    } else {
+        src.advance(mem::size_of::<i32>());
+    }
+
+    if fields.contains(Fields::ALIGNMENT_START) {
+        *record.position_mut() = get_position(src)?;
+    } else {
+        src.advance(mem::size_of::<i32>());
+    }
 
     let l_read_name = NonZeroUsize::new(usize::from(src.get_u8()))
         .ok_or_else(|| io::Error::new(io::ErrorKind::InvalidData, "invalid l_read_name"))?;
 
-    *record.mapping_quality_mut() = get_mapping_quality(src)?;
+    if fields.contains(Fields::MAPPING_QUALITY) {
+        *record.mapping_quality_mut() = get_mapping_quality(src)?;
+    } else {
+        src.advance(mem::size_of::<u8>());
+    }
 
     // Discard bin.
     src.advance(mem::size_of::<u16>());
 
     let n_cigar_op = usize::from(src.get_u16_le());
 
-    *record.flags_mut() = get_flags(src)?;
+    if fields.contains(Fields::FLAGS) {
+        *record.flags_mut() = get_flags(src)?;
+    } else {
+        src.advance(mem::size_of::<u16>());
+    }
 
     let l_seq = usize::try_from(src.get_u32_le())
         .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
 
-    *record.mate_reference_sequence_id_mut() = get_reference_sequence_id(src)?;
-    *record.mate_position_mut() = get_position(src)?;
+    if fields.contains(Fields::MATE_REFERENCE_SEQUENCE_ID) {
+        *record.mate_reference_sequence_id_mut() = get_reference_sequence_id(src)?;
+    } else {
+        src.advance(mem::size_of::<i32>());
+    }
 
-    *record.template_length_mut() = src.get_i32_le();
+    if fields.contains(Fields::MATE_ALIGNMENT_START) {
+        *record.mate_position_mut() = get_position(src)?;
+    } else {
+        src.advance(mem::size_of::<i32>());
+    }
 
-    get_read_name(src, record.read_name_mut(), l_read_name)?;
-    get_cigar(src, record.cigar_mut(), n_cigar_op)?;
-    get_sequence(src, record.sequence_mut(), l_seq)?;
-    get_quality_scores(src, record.quality_scores_mut(), l_seq)?;
+    if fields.contains(Fields::TEMPLATE_LENGTH) {
+        *record.template_length_mut() = src.get_i32_le();
+    } else {
+        src.advance(mem::size_of::<i32>());
+    }
 
-    get_data(src, record.data_mut())?;
+    if fields.contains(Fields::READ_NAME) {
+        get_read_name(src, record.read_name_mut(), l_read_name)?;
+    } else {
+        src.advance(usize::from(l_read_name));
+    }
+
+    if fields.contains(Fields::CIGAR) {
+        get_cigar(src, record.cigar_mut(), n_cigar_op)?;
+    } else {
+        let len = mem::size_of::<u32>() * n_cigar_op;
+        src.advance(len);
+    }
+
+    if fields.contains(Fields::SEQUENCE) {
+        get_sequence(src, record.sequence_mut(), l_seq)?;
+    } else {
+        let len = (l_seq + 1) / 2;
+        src.advance(len);
+    }
+
+    if fields.contains(Fields::QUALITY_SCORES) {
+        get_quality_scores(src, record.quality_scores_mut(), l_seq)?;
+    } else {
+        src.advance(l_seq);
+    }
+
+    if fields.contains(Fields::DATA) {
+        get_data(src, record.data_mut())?;
+    }
 
     Ok(())
 }