+++ noodles-sam/src/reader/record.rs
@@ -1,0 +1,5 @@
+//! SAM record reader.
+
+mod fields;
+
+pub use self::fields::Fields;